xds: do not attempt to push resources to Ztunnel (#54380)
This is impossible in standard operations. However, a client can connect
to Istiod with arbitrary XDS info. If they declare themselves `ztunnel`,
we don't compute SidecarScope. Most of the generators make an assumption
SidecarScope is non-nil and don't check it. To handle this, we skip
these at the top level.